refactored translations to a seperate directory
