+++ TrueHue/assets/translations/textTranslations.js
@@ -1,0 +1,247 @@
+// Create translations for this screen
+const screenTranslations = {
+  en: {
+    appTitle: "Wood Analysis Assistant",
+    instruction:
+      "Upload or take a photo of wood to analyze its type and properties",
+    chooseGallery: "Choose from Gallery",
+    takePhoto: "Take a Photo",
+    analyzing: "Analyzing wood type...",
+    generatingReport: "Generating comprehensive report...",
+    selectWoodPrompt: "Select wood type for analysis:",
+    mediumCherry: "Medium Cherry",
+    desertOak: "Desert Oak",
+    graphiteWalnut: "Graphite Walnut",
+    cancel: "Cancel",
+    analyze: "Analyze",
+    generateReport: "Generate Report",
+    newImage: "New Image",
+    colorAnalysis: "Color Analysis",
+    dark: "Dark",
+    light: "Light",
+    analysis: "Analysis",
+    confidence: "Confidence",
+    analysisResults: "Analysis Results",
+    reportTitle: "Comprehensive Analysis",
+    woodClassification: "Wood Classification",
+    unknown: "Unknown",
+    allProbabilities: "All Probabilities:",
+    whatDoesThisMean: "What does this mean?",
+    saveReport: "Save Report",
+    shareResults: "Share Results",
+    noImage: "No image selected",
+    selectFirst: "Please select an image first.",
+    reportGenerated: "Report Generated",
+    seeDetails: "See below for full details",
+    reportSaved: "Report saved successfully!",
+    errorSaving: "Failed to save report.",
+    woodValidation: "Wood Validation",
+    binaryClassification: "Binary Classification",
+    multiclassClassification: "Multiclass Classification",
+    regressionAnalysis: "Regression Analysis",
+    validationClassification: "Validation Classification",
+    result: "Result",
+    value: "Value",
+    analysisNotification: "Analysis Complete",
+    analysisNotificationBody:
+      "Your wood analysis has been completed successfully.",
+    reportNotification: "Report Generated",
+    reportNotificationBody:
+      "Your comprehensive wood report is now ready to view.",
+  },
+  es: {
+    appTitle: "Asistente de análisis de madera",
+    instruction:
+      "Sube o toma una foto de la madera para analizar su tipo y propiedades",
+    chooseGallery: "Elegir de la galería",
+    takePhoto: "Tomar una foto",
+    analyzing: "Analizando tipo de madera...",
+    generatingReport: "Generando informe completo...",
+    selectWoodPrompt: "Seleccione tipo de madera para análisis:",
+    mediumCherry: "Cerezo Medio",
+    desertOak: "Roble Desierto",
+    graphiteWalnut: "Nogal Grafito",
+    cancel: "Cancelar",
+    analyze: "Analizar",
+    generateReport: "Generar informe",
+    newImage: "Nueva imagen",
+    colorAnalysis: "Análisis de color",
+    dark: "Oscuro",
+    light: "Claro",
+    analysis: "Análisis",
+    confidence: "Confianza",
+    analysisResults: "Resultados del análisis",
+    reportTitle: "Análisis completo",
+    woodClassification: "Clasificación de madera",
+    unknown: "Desconocido",
+    allProbabilities: "Todas las probabilidades:",
+    whatDoesThisMean: "¿Qué significa esto?",
+    saveReport: "Guardar informe",
+    shareResults: "Compartir resultados",
+    noImage: "No se ha seleccionado imagen",
+    selectFirst: "Por favor, seleccione una imagen primero.",
+    reportGenerated: "Informe generado",
+    seeDetails: "Ver detalles completos abajo",
+    reportSaved: "¡Informe guardado con éxito!",
+    errorSaving: "Error al guardar el informe.",
+    woodValidation: "Validación de madera",
+    binaryClassification: "Clasificación binaria",
+    multiclassClassification: "Clasificación multiclase",
+    regressionAnalysis: "Análisis de regresión",
+    validationClassification: "Clasificación de validación",
+    result: "Resultado",
+    value: "Valor",
+    analysisNotification: "Análisis Completado",
+    analysisNotificationBody:
+      "Su análisis de madera se ha completado con éxito.",
+    reportNotification: "Informe Generado",
+    reportNotificationBody:
+      "Su informe completo de madera está listo para ver.",
+  },
+  fr: {
+    appTitle: "Assistant d'analyse du bois",
+    instruction:
+      "Téléchargez ou prenez une photo du bois pour analyser son type et ses propriétés",
+    chooseGallery: "Choisir dans la galerie",
+    takePhoto: "Prendre une photo",
+    analyzing: "Analyse du type de bois...",
+    generatingReport: "Génération d'un rapport complet...",
+    selectWoodPrompt: "Sélectionnez le type de bois à analyser :",
+    mediumCherry: "Cerisier Moyen",
+    desertOak: "Chêne Désert",
+    graphiteWalnut: "Noyer Graphite",
+    cancel: "Annuler",
+    analyze: "Analyser",
+    generateReport: "Générer un rapport",
+    newImage: "Nouvelle image",
+    colorAnalysis: "Analyse des couleurs",
+    dark: "Foncé",
+    light: "Clair",
+    analysis: "Analyse",
+    confidence: "Confiance",
+    analysisResults: "Résultats de l'analyse",
+    reportTitle: "Analyse complète",
+    woodClassification: "Classification du bois",
+    unknown: "Inconnu",
+    allProbabilities: "Toutes les probabilités :",
+    whatDoesThisMean: "Qu'est-ce que cela signifie ?",
+    saveReport: "Enregistrer le rapport",
+    shareResults: "Partager les résultats",
+    noImage: "Aucune image sélectionnée",
+    selectFirst: "Veuillez d'abord sélectionner une image.",
+    reportGenerated: "Rapport généré",
+    seeDetails: "Voir les détails complets ci-dessous",
+    reportSaved: "Rapport enregistré avec succès !",
+    errorSaving: "Échec de l'enregistrement du rapport.",
+    woodValidation: "Validation du bois",
+    binaryClassification: "Classification binaire",
+    multiclassClassification: "Classification multiclasse",
+    regressionAnalysis: "Analyse de régression",
+    validationClassification: "Classification de validation",
+    result: "Résultat",
+    value: "Valeur",
+    analysisNotification: "Analyse Terminée",
+    analysisNotificationBody:
+      "Votre analyse de bois a été complétée avec succès.",
+    reportNotification: "Rapport Généré",
+    reportNotificationBody:
+      "Votre rapport complet sur le bois est maintenant prêt à être consulté.",
+  },
+  de: {
+    appTitle: "Holzanalyse-Assistent",
+    instruction:
+      "Laden Sie ein Foto hoch oder machen Sie ein Foto von Holz, um seinen Typ und seine Eigenschaften zu analysieren",
+    chooseGallery: "Aus Galerie wählen",
+    takePhoto: "Foto aufnehmen",
+    analyzing: "Holztyp wird analysiert...",
+    generatingReport: "Umfassender Bericht wird erstellt...",
+    selectWoodPrompt: "Wählen Sie Holztyp für die Analyse:",
+    mediumCherry: "Mittlere Kirsche",
+    desertOak: "Wüsteneiche",
+    graphiteWalnut: "Graphit-Walnuss",
+    cancel: "Abbrechen",
+    analyze: "Analysieren",
+    generateReport: "Bericht erstellen",
+    newImage: "Neues Bild",
+    colorAnalysis: "Farbanalyse",
+    dark: "Dunkel",
+    light: "Hell",
+    analysis: "Analyse",
+    confidence: "Konfidenz",
+    analysisResults: "Analyseergebnisse",
+    reportTitle: "Umfassende Analyse",
+    woodClassification: "Holzklassifizierung",
+    unknown: "Unbekannt",
+    allProbabilities: "Alle Wahrscheinlichkeiten:",
+    whatDoesThisMean: "Was bedeutet das?",
+    saveReport: "Bericht speichern",
+    shareResults: "Ergebnisse teilen",
+    noImage: "Kein Bild ausgewählt",
+    selectFirst: "Bitte wählen Sie zuerst ein Bild aus.",
+    reportGenerated: "Bericht erstellt",
+    seeDetails: "Vollständige Details unten anzeigen",
+    reportSaved: "Bericht erfolgreich gespeichert!",
+    errorSaving: "Fehler beim Speichern des Berichts.",
+    woodValidation: "Holzvalidierung",
+    binaryClassification: "Binäre Klassifizierung",
+    multiclassClassification: "Mehrklassen-Klassifizierung",
+    regressionAnalysis: "Regressionsanalyse",
+    validationClassification: "Validierungsklassifizierung",
+    result: "Ergebnis",
+    value: "Wert",
+    analysisNotification: "Analyse Abgeschlossen",
+    analysisNotificationBody:
+      "Ihre Holzanalyse wurde erfolgreich abgeschlossen.",
+    reportNotification: "Bericht Erstellt",
+    reportNotificationBody:
+      "Ihr umfassender Holzbericht ist jetzt zur Ansicht bereit.",
+  },
+  zh: {
+    appTitle: "木材分析助手",
+    instruction: "上传或拍摄木材照片，分析其类型和特性",
+    chooseGallery: "从相册选择",
+    takePhoto: "拍照",
+    analyzing: "正在分析木材类型...",
+    generatingReport: "正在生成综合报告...",
+    selectWoodPrompt: "选择木材类型进行分析：",
+    mediumCherry: "中等樱桃木",
+    desertOak: "沙漠橡木",
+    graphiteWalnut: "石墨胡桃木",
+    cancel: "取消",
+    analyze: "分析",
+    generateReport: "生成报告",
+    newImage: "新图像",
+    colorAnalysis: "颜色分析",
+    dark: "深色",
+    light: "浅色",
+    analysis: "分析",
+    confidence: "置信度",
+    analysisResults: "分析结果",
+    reportTitle: "综合分析",
+    woodClassification: "木材分类",
+    unknown: "未知",
+    allProbabilities: "所有概率：",
+    whatDoesThisMean: "这意味着什么？",
+    saveReport: "保存报告",
+    shareResults: "分享结果",
+    noImage: "未选择图像",
+    selectFirst: "请先选择一张图像。",
+    reportGenerated: "报告已生成",
+    seeDetails: "查看以下完整详情",
+    reportSaved: "报告保存成功！",
+    errorSaving: "保存报告失败。",
+    woodValidation: "木材验证",
+    binaryClassification: "二元分类",
+    multiclassClassification: "多类分类",
+    regressionAnalysis: "回归分析",
+    validationClassification: "验证分类",
+    result: "结果",
+    value: "值",
+    analysisNotification: "分析完成",
+    analysisNotificationBody: "您的木材分析已成功完成。",
+    reportNotification: "报告已生成",
+    reportNotificationBody: "您的综合木材报告现已准备好查看。",
+  },
+};
+
+export default screenTranslations;
