+++ TrueHue/assets/translations/textTranslationsIndex.js
@@ -1,0 +1,227 @@
+// Export translations for all supported languages
+export const translations = {
+  en: {
+    settings: "Settings",
+    language: "Language",
+    languageDesc: "Select your preferred language for the app interface",
+    notifications: "Notifications",
+    notificationsDesc: "Manage your notification preferences",
+    enableNotifications: "Enable Notifications",
+    notificationsToggleDesc: "Turn all notifications on or off",
+    analysisResults: "Analysis Results",
+    analysisResultsDesc:
+      "Receive notifications when wood analyses are complete",
+    reportGeneration: "Report Generation",
+    reportGenerationDesc: "Receive notifications when reports are ready",
+    appearance: "Appearance",
+    appearanceDesc: "Customize the visual appearance of the app",
+    darkMode: "Dark Mode",
+    darkModeDesc: "Use dark colors for the app interface",
+    storageData: "Storage & Data",
+    storageDataDesc: "Manage how your settings and data are saved",
+    autoSave: "Auto-Save Settings",
+    autoSaveDesc: "Automatically save settings when changed",
+    saveSettings: "Save Settings",
+    saving: "Saving...",
+    resetDefaults: "Reset to Defaults",
+    support: "Support",
+    supportDesc: "Get help with the app or report issues",
+    helpFaq: "Help & FAQ",
+    contactSupport: "Contact Support",
+    reportBug: "Report a Bug",
+    versionInfo: "Wood Analysis App",
+    version: "Version 1.0.0",
+    resetSettingsTitle: "Reset Settings",
+    resetSettingsDesc:
+      "Are you sure you want to reset all settings to default values?",
+    cancel: "Cancel",
+    reset: "Reset",
+    success: "Success",
+    settingsSaved: "Your settings have been saved successfully.",
+    error: "Error",
+    settingsSaveError: "Failed to save settings. Please try again.",
+    loadingSettings: "Loading settings...",
+  },
+  es: {
+    settings: "Configuración",
+    language: "Idioma",
+    languageDesc:
+      "Seleccione su idioma preferido para la interfaz de la aplicación",
+    notifications: "Notificaciones",
+    notificationsDesc: "Administre sus preferencias de notificación",
+    enableNotifications: "Habilitar notificaciones",
+    notificationsToggleDesc: "Activar o desactivar todas las notificaciones",
+    analysisResults: "Resultados de análisis",
+    analysisResultsDesc:
+      "Recibir notificaciones cuando se completen los análisis de madera",
+    reportGeneration: "Generación de informes",
+    reportGenerationDesc:
+      "Recibir notificaciones cuando los informes estén listos",
+    appearance: "Apariencia",
+    appearanceDesc: "Personalizar la apariencia visual de la aplicación",
+    darkMode: "Modo oscuro",
+    darkModeDesc: "Usar colores oscuros para la interfaz de la aplicación",
+    storageData: "Almacenamiento y datos",
+    storageDataDesc: "Administre cómo se guardan sus configuraciones y datos",
+    autoSave: "Guardar automáticamente",
+    autoSaveDesc: "Guardar automáticamente la configuración cuando cambie",
+    saveSettings: "Guardar configuración",
+    saving: "Guardando...",
+    resetDefaults: "Restablecer valores predeterminados",
+    support: "Soporte",
+    supportDesc: "Obtenga ayuda con la aplicación o informe problemas",
+    helpFaq: "Ayuda y preguntas frecuentes",
+    contactSupport: "Contactar con soporte",
+    reportBug: "Informar de un error",
+    versionInfo: "Aplicación de análisis de madera",
+    version: "Versión 1.0.0",
+    resetSettingsTitle: "Restablecer configuración",
+    resetSettingsDesc:
+      "¿Está seguro de que desea restablecer toda la configuración a los valores predeterminados?",
+    cancel: "Cancelar",
+    reset: "Restablecer",
+    success: "Éxito",
+    settingsSaved: "Su configuración se ha guardado correctamente.",
+    error: "Error",
+    settingsSaveError:
+      "No se pudo guardar la configuración. Inténtalo de nuevo.",
+    loadingSettings: "Cargando configuración...",
+  },
+  fr: {
+    settings: "Paramètres",
+    language: "Langue",
+    languageDesc:
+      "Sélectionnez votre langue préférée pour l'interface de l'application",
+    notifications: "Notifications",
+    notificationsDesc: "Gérez vos préférences de notification",
+    enableNotifications: "Activer les notifications",
+    notificationsToggleDesc: "Activer ou désactiver toutes les notifications",
+    analysisResults: "Résultats d'analyse",
+    analysisResultsDesc:
+      "Recevoir des notifications lorsque les analyses du bois sont terminées",
+    reportGeneration: "Génération de rapports",
+    reportGenerationDesc:
+      "Recevoir des notifications lorsque les rapports sont prêts",
+    appearance: "Apparence",
+    appearanceDesc: "Personnalisez l'apparence visuelle de l'application",
+    darkMode: "Mode sombre",
+    darkModeDesc:
+      "Utiliser des couleurs sombres pour l'interface de l'application",
+    storageData: "Stockage et données",
+    storageDataDesc:
+      "Gérez la façon dont vos paramètres et données sont enregistrés",
+    autoSave: "Enregistrement automatique",
+    autoSaveDesc:
+      "Enregistrer automatiquement les paramètres lorsqu'ils sont modifiés",
+    saveSettings: "Enregistrer les paramètres",
+    saving: "Enregistrement...",
+    resetDefaults: "Réinitialiser aux valeurs par défaut",
+    support: "Support",
+    supportDesc:
+      "Obtenez de l'aide avec l'application ou signalez des problèmes",
+    helpFaq: "Aide et FAQ",
+    contactSupport: "Contacter le support",
+    reportBug: "Signaler un bug",
+    versionInfo: "Application d'analyse du bois",
+    version: "Version 1.0.0",
+    resetSettingsTitle: "Réinitialiser les paramètres",
+    resetSettingsDesc:
+      "Êtes-vous sûr de vouloir réinitialiser tous les paramètres aux valeurs par défaut?",
+    cancel: "Annuler",
+    reset: "Réinitialiser",
+    success: "Succès",
+    settingsSaved: "Vos paramètres ont été enregistrés avec succès.",
+    error: "Erreur",
+    settingsSaveError:
+      "Échec de l'enregistrement des paramètres. Veuillez réessayer.",
+    loadingSettings: "Chargement des paramètres...",
+  },
+  de: {
+    settings: "Einstellungen",
+    language: "Sprache",
+    languageDesc: "Wählen Sie Ihre bevorzugte Sprache für die App-Oberfläche",
+    notifications: "Benachrichtigungen",
+    notificationsDesc: "Verwalten Sie Ihre Benachrichtigungseinstellungen",
+    enableNotifications: "Benachrichtigungen aktivieren",
+    notificationsToggleDesc: "Alle Benachrichtigungen ein- oder ausschalten",
+    analysisResults: "Analyseergebnisse",
+    analysisResultsDesc:
+      "Benachrichtigungen erhalten, wenn Holzanalysen abgeschlossen sind",
+    reportGeneration: "Berichterstellung",
+    reportGenerationDesc:
+      "Benachrichtigungen erhalten, wenn Berichte fertig sind",
+    appearance: "Erscheinungsbild",
+    appearanceDesc: "Passen Sie das visuelle Erscheinungsbild der App an",
+    darkMode: "Dunkler Modus",
+    darkModeDesc: "Dunkle Farben für die App-Oberfläche verwenden",
+    storageData: "Speicher & Daten",
+    storageDataDesc:
+      "Verwalten Sie, wie Ihre Einstellungen und Daten gespeichert werden",
+    autoSave: "Automatisches Speichern",
+    autoSaveDesc:
+      "Einstellungen automatisch speichern, wenn sie geändert werden",
+    saveSettings: "Einstellungen speichern",
+    saving: "Speichern...",
+    resetDefaults: "Auf Standardwerte zurücksetzen",
+    support: "Unterstützung",
+    supportDesc: "Erhalten Sie Hilfe zur App oder melden Sie Probleme",
+    helpFaq: "Hilfe & FAQ",
+    contactSupport: "Support kontaktieren",
+    reportBug: "Fehler melden",
+    versionInfo: "Holzanalyse-App",
+    version: "Version 1.0.0",
+    resetSettingsTitle: "Einstellungen zurücksetzen",
+    resetSettingsDesc:
+      "Sind Sie sicher, dass Sie alle Einstellungen auf die Standardwerte zurücksetzen möchten?",
+    cancel: "Abbrechen",
+    reset: "Zurücksetzen",
+    success: "Erfolg",
+    settingsSaved: "Ihre Einstellungen wurden erfolgreich gespeichert.",
+    error: "Fehler",
+    settingsSaveError:
+      "Einstellungen konnten nicht gespeichert werden. Bitte versuchen Sie es erneut.",
+    loadingSettings: "Einstellungen werden geladen...",
+  },
+  zh: {
+    settings: "设置",
+    language: "语言",
+    languageDesc: "选择您偏好的应用界面语言",
+    notifications: "通知",
+    notificationsDesc: "管理您的通知首选项",
+    enableNotifications: "启用通知",
+    notificationsToggleDesc: "打开或关闭所有通知",
+    analysisResults: "分析结果",
+    analysisResultsDesc: "木材分析完成时接收通知",
+    reportGeneration: "报告生成",
+    reportGenerationDesc: "报告准备就绪时接收通知",
+    appearance: "外观",
+    appearanceDesc: "自定义应用程序的视觉外观",
+    darkMode: "深色模式",
+    darkModeDesc: "为应用界面使用深色颜色",
+    storageData: "存储和数据",
+    storageDataDesc: "管理您的设置和数据的保存方式",
+    autoSave: "自动保存设置",
+    autoSaveDesc: "更改设置时自动保存",
+    saveSettings: "保存设置",
+    saving: "保存中...",
+    resetDefaults: "重置为默认值",
+    support: "支持",
+    supportDesc: "获取应用帮助或报告问题",
+    helpFaq: "帮助和常见问题",
+    contactSupport: "联系支持",
+    reportBug: "报告错误",
+    versionInfo: "木材分析应用",
+    version: "版本 1.0.0",
+    resetSettingsTitle: "重置设置",
+    resetSettingsDesc: "您确定要将所有设置重置为默认值吗？",
+    cancel: "取消",
+    reset: "重置",
+    success: "成功",
+    settingsSaved: "您的设置已成功保存。",
+    error: "错误",
+    settingsSaveError: "无法保存设置。请再试一次。",
+    loadingSettings: "加载设置中...",
+  },
+};
+
+export default translations;
