+++ TrueHue/assets/translations/textTranslationsRgb.js
@@ -1,0 +1,172 @@
+// Create translations for this screen
+const screenTranslations = {
+  en: {
+    appTitle: "Wood Veneer Comparison",
+    instruction: "Compare two veneer samples to see how similar they are",
+    veneerSample1: "Veneer Sample 1",
+    veneerSample2: "Veneer Sample 2",
+    chooseGallery: "Choose from Gallery",
+    takePhoto: "Take a Photo",
+    changeImage: "Change Image",
+    calculateDifference: "Calculate Difference",
+    calculating: "Calculating...",
+    reset: "Reset",
+    comparisonResults: "Comparison Results",
+    rgbDifference: "RGB Euclidean Difference",
+    similar: "Similar",
+    different: "Different",
+    verySimilar: "These veneer samples are very similar in color.",
+    moderateDifference: "These veneer samples have moderate color differences.",
+    significantDifference:
+      "These veneer samples have significant color differences.",
+    whatDoesItMean: "What does this mean?",
+    explanation:
+      "The RGB Euclidean difference measures the distance between colors in RGB space. A lower value indicates more similar colors between the veneer samples. This can help determine if two wood samples will match visually when placed together.",
+    noGalleryPermission: "Permission for media access not granted.",
+    noCameraPermission: "Permission for camera access is not granted.",
+    missingImages: "Missing Images",
+    selectBothImages: "Please select both images to compare.",
+    error: "Error",
+    failedCalculate: "Failed to calculate difference: ",
+    unexpectedError: "An unexpected error occurred",
+    analysisNotification: "Comparison Complete",
+    analysisNotificationBody:
+      "Your wood veneer comparison has been completed successfully.",
+  },
+  es: {
+    appTitle: "Comparación de Chapas de Madera",
+    instruction: "Compare dos muestras de chapa para ver qué tan similares son",
+    veneerSample1: "Muestra de Chapa 1",
+    veneerSample2: "Muestra de Chapa 2",
+    chooseGallery: "Elegir de la Galería",
+    takePhoto: "Tomar una Foto",
+    changeImage: "Cambiar Imagen",
+    calculateDifference: "Calcular Diferencia",
+    calculating: "Calculando...",
+    reset: "Reiniciar",
+    comparisonResults: "Resultados de Comparación",
+    rgbDifference: "Diferencia Euclidiana RGB",
+    similar: "Similar",
+    different: "Diferente",
+    verySimilar: "Estas muestras de chapa son muy similares en color.",
+    moderateDifference:
+      "Estas muestras de chapa tienen diferencias moderadas de color.",
+    significantDifference:
+      "Estas muestras de chapa tienen diferencias significativas de color.",
+    whatDoesItMean: "¿Qué significa esto?",
+    explanation:
+      "La diferencia euclidiana RGB mide la distancia entre colores en el espacio RGB. Un valor más bajo indica colores más similares entre las muestras de chapa. Esto puede ayudar a determinar si dos muestras de madera coincidirán visualmente cuando se coloquen juntas.",
+    noGalleryPermission: "Permiso para acceso a medios no concedido.",
+    noCameraPermission: "Permiso para acceso a la cámara no concedido.",
+    missingImages: "Imágenes Faltantes",
+    selectBothImages: "Por favor seleccione ambas imágenes para comparar.",
+    error: "Error",
+    failedCalculate: "Error al calcular la diferencia: ",
+    unexpectedError: "Ocurrió un error inesperado",
+    analysisNotification: "Comparación Completada",
+    analysisNotificationBody:
+      "Su comparación de chapas de madera se ha completado con éxito.",
+  },
+  fr: {
+    appTitle: "Comparaison de Placages de Bois",
+    instruction:
+      "Comparez deux échantillons de placage pour voir à quel point ils sont similaires",
+    veneerSample1: "Échantillon de Placage 1",
+    veneerSample2: "Échantillon de Placage 2",
+    chooseGallery: "Choisir dans la Galerie",
+    takePhoto: "Prendre une Photo",
+    changeImage: "Changer l'Image",
+    calculateDifference: "Calculer la Différence",
+    calculating: "Calcul en cours...",
+    reset: "Réinitialiser",
+    comparisonResults: "Résultats de Comparaison",
+    rgbDifference: "Différence Euclidienne RGB",
+    similar: "Similaire",
+    different: "Différent",
+    verySimilar: "Ces échantillons de placage sont très similaires en couleur.",
+    moderateDifference:
+      "Ces échantillons de placage présentent des différences de couleur modérées.",
+    significantDifference:
+      "Ces échantillons de placage présentent des différences de couleur significatives.",
+    whatDoesItMean: "Qu'est-ce que cela signifie ?",
+    explanation:
+      "La différence euclidienne RGB mesure la distance entre les couleurs dans l'espace RGB. Une valeur plus basse indique des couleurs plus similaires entre les échantillons de placage. Cela peut aider à déterminer si deux échantillons de bois s'accorderont visuellement lorsqu'ils seront placés ensemble.",
+    noGalleryPermission: "Permission d'accès aux médias non accordée.",
+    noCameraPermission: "Permission d'accès à la caméra non accordée.",
+    missingImages: "Images Manquantes",
+    selectBothImages: "Veuillez sélectionner les deux images à comparer.",
+    error: "Erreur",
+    failedCalculate: "Échec du calcul de la différence : ",
+    unexpectedError: "Une erreur inattendue s'est produite",
+    analysisNotification: "Comparaison Terminée",
+    analysisNotificationBody:
+      "Votre comparaison de placages de bois a été complétée avec succès.",
+  },
+  de: {
+    appTitle: "Holzfurnier-Vergleich",
+    instruction:
+      "Vergleichen Sie zwei Furnierproben, um zu sehen, wie ähnlich sie sind",
+    veneerSample1: "Furnierprobe 1",
+    veneerSample2: "Furnierprobe 2",
+    chooseGallery: "Aus Galerie wählen",
+    takePhoto: "Foto aufnehmen",
+    changeImage: "Bild ändern",
+    calculateDifference: "Unterschied berechnen",
+    calculating: "Berechne...",
+    reset: "Zurücksetzen",
+    comparisonResults: "Vergleichsergebnisse",
+    rgbDifference: "RGB Euklidische Differenz",
+    similar: "Ähnlich",
+    different: "Unterschiedlich",
+    verySimilar: "Diese Furnierproben haben sehr ähnliche Farben.",
+    moderateDifference: "Diese Furnierproben haben mäßige Farbunterschiede.",
+    significantDifference:
+      "Diese Furnierproben haben signifikante Farbunterschiede.",
+    whatDoesItMean: "Was bedeutet das?",
+    explanation:
+      "Die RGB Euklidische Differenz misst den Abstand zwischen Farben im RGB-Raum. Ein niedrigerer Wert zeigt ähnlichere Farben zwischen den Furnierproben an. Dies kann helfen zu bestimmen, ob zwei Holzproben visuell zusammenpassen, wenn sie nebeneinander platziert werden.",
+    noGalleryPermission: "Berechtigung für Medienzugriff nicht erteilt.",
+    noCameraPermission: "Berechtigung für Kamerazugriff nicht erteilt.",
+    missingImages: "Fehlende Bilder",
+    selectBothImages: "Bitte wählen Sie beide Bilder zum Vergleichen aus.",
+    error: "Fehler",
+    failedCalculate: "Berechnung des Unterschieds fehlgeschlagen: ",
+    unexpectedError: "Ein unerwarteter Fehler ist aufgetreten",
+    analysisNotification: "Vergleich Abgeschlossen",
+    analysisNotificationBody:
+      "Ihr Holzfurnier-Vergleich wurde erfolgreich abgeschlossen.",
+  },
+  zh: {
+    appTitle: "木质贴面比较",
+    instruction: "比较两个贴面样本，看看它们有多相似",
+    veneerSample1: "贴面样本 1",
+    veneerSample2: "贴面样本 2",
+    chooseGallery: "从相册选择",
+    takePhoto: "拍照",
+    changeImage: "更换图片",
+    calculateDifference: "计算差异",
+    calculating: "计算中...",
+    reset: "重置",
+    comparisonResults: "比较结果",
+    rgbDifference: "RGB欧几里得差异",
+    similar: "相似",
+    different: "不同",
+    verySimilar: "这些贴面样本在颜色上非常相似。",
+    moderateDifference: "这些贴面样本有中等的颜色差异。",
+    significantDifference: "这些贴面样本有显著的颜色差异。",
+    whatDoesItMean: "这意味着什么？",
+    explanation:
+      "RGB欧几里得差异测量RGB空间中颜色之间的距离。较低的值表示贴面样本之间的颜色更相似。这可以帮助确定两个木材样本放在一起时是否在视觉上匹配。",
+    noGalleryPermission: "未授予媒体访问权限。",
+    noCameraPermission: "未授予相机访问权限。",
+    missingImages: "缺少图像",
+    selectBothImages: "请选择两个图像进行比较。",
+    error: "错误",
+    failedCalculate: "计算差异失败: ",
+    unexpectedError: "发生意外错误",
+    analysisNotification: "比较完成",
+    analysisNotificationBody: "您的木质贴面比较已成功完成。",
+  },
+};
+
+export default screenTranslations;
